third_party/WebKit/LayoutTests/http/tests/serviceworker/unregister-then-register-new-script.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var worker_url = 'resources/empty-worker.js';

async_test(function(t) {
    var scope = 'scope/register-waits-for-unregistered-registration-to-clear';
    var new_worker_url = worker_url + '?new';
    var iframe;
    var registration;
    var unloaded = false;

    service_worker_unregister_and_register(t, worker_url, scope)
      .then(function(r) {
          registration = r;
          return wait_for_update(t, registration);
        })
      .then(function(worker) {
          return wait_for_state(t, worker, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          iframe = frame;
          return registration.unregister();
        })
      .then(function() {
          setTimeout(function() {
              unloaded = true;
              unload_iframe(iframe);
            }, 10);
          return navigator.serviceWorker.register(new_worker_url,
                                                  { scope: scope });
        })
      .then(function(new_registration) {
          assert_true(unloaded,
                      'register should not resolve until iframe unloaded');
          assert_equals(registration.installing, null,
                        'registration.installing');
          assert_equals(registration.waiting, null, 'registration.waiting');
          assert_equals(registration.active, null, 'registration.active');
          return new_registration.unregister();
        })
      .then(function() {
          t.done();
        })
      .catch(unreached_rejection(t));
  }, 'Registering a new script URL does not resolve until unregistered ' +
     'registration is cleared');

async_test(function(t) {
    var scope = 'scope/unregister-then-register-new-script-that-exists';
    var new_worker_url = worker_url + '?new';
    var iframe;
    var registration;

    service_worker_unregister_and_register(t, worker_url, scope)
      .then(function(r) {
          registration = r;
          return wait_for_update(t, registration);
        })
      .then(function(worker) {
          return wait_for_state(t, worker, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          iframe = frame;
          return registration.unregister();
        })
      .then(function() {
          var promise = navigator.serviceWorker.register(new_worker_url,
                                                         { scope: scope });
          unload_iframe(iframe);
          return promise;
        })
      .then(function(new_registration) {
          assert_not_equals(registration, new_registration,
                            'register() should resolve to a new registration');
          assert_equals(registration.installing, null,
                        'old registration.installing');
          assert_equals(registration.waiting, null,
                        'old registration.waiting');
          assert_equals(registration.active, null,
                       'old registration.active');
          registration = new_registration;
          return wait_for_update(t, registration);
        })
      .then(function(worker) {
          assert_equals(registration.installing.scriptURL,
                        normalizeURL(new_worker_url),
                        'new registration.installing');
          assert_equals(registration.waiting, null,
                        'new registration.waiting');
          assert_equals(registration.active, null,
                       'new registration.active');
          return wait_for_state(t, worker, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          assert_equals(
              frame.contentWindow.navigator.serviceWorker.controller.scriptURL,
              normalizeURL(new_worker_url),
              'the new worker should control a new document');
          unload_iframe(frame);
          return registration.unregister();
        })
      .then(function() {
          t.done();
        })
      .catch(unreached_rejection(t));
}, 'Registering a new script URL while an unregistered registration is in use');

async_test(function(t) {
    var scope = 'scope/unregister-then-register-new-script-that-404s';
    var iframe;
    var registration;

    service_worker_unregister_and_register(t, worker_url, scope)
      .then(function(r) {
          registration = r;
          return wait_for_update(t, registration);
        })
      .then(function(worker) {
          return wait_for_state(t, worker, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          iframe = frame;
          return registration.unregister();
        })
      .then(function() {
          var promise = navigator.serviceWorker.register('this-will-404',
                                                         { scope: scope });
          unload_iframe(iframe);
          return promise;
        })
      .then(
        function() {
          assert_unreached('register should reject the promise');
        },
        function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          assert_equals(frame.contentWindow.navigator.serviceWorker.controller,
                        null,
                        'document should not load with a controller');
          unload_iframe(frame);
          t.done();
        })
      .catch(unreached_rejection(t));
}, 'Registering a new script URL that 404s does not resurrect an ' +
       'unregistered registration');

async_test(function(t) {
    var scope = 'scope/unregister-then-register-reject-install-worker';
    var iframe;
    var registration;

    service_worker_unregister_and_register(t, worker_url, scope)
      .then(function(r) {
          registration = r;
          return wait_for_update(t, registration);
        })
      .then(function(worker) {
          return wait_for_state(t, worker, 'activated');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          iframe = frame;
          return registration.unregister();
        })
      .then(function() {
          var promise = navigator.serviceWorker.register(
              'resources/reject-install-worker.js', { scope: scope });
          unload_iframe(iframe);
          return promise;
        })
      .then(function(r) {
          registration = r;
          return wait_for_update(t, registration);
        })
      .then(function(worker) {
          return wait_for_state(t, worker, 'redundant');
        })
      .then(function() {
          return with_iframe(scope);
        })
      .then(function(frame) {
          assert_equals(frame.contentWindow.navigator.serviceWorker.controller,
                        null,
                        'document should not load with a controller');
          unload_iframe(frame);
          return registration.unregister();
        })
      .then(function() {
          t.done();
        })
      .catch(unreached_rejection(t));
  }, 'Registering a new script URL that fails to install does not resurrect ' +
       'an unregistered registration');
</script>
